Added tests for update and get recommendation
