+++ backend-spring-boot/src/test/java/csd/product/ProductIntegrationTest.java
@@ -31,9 +31,6 @@ import csd.app.user.User;
 import csd.app.user.UserRepository;
 import csd.app.user.UserInfo;
 import csd.app.user.UserInfoRepository;
-import csd.app.user.UserRecommendation;
-import csd.app.user.UserRecommendationRepository;
-import csd.app.payload.request.RecommendationRequest;
 
 /** Start an actual HTTP server listening at a random port */
 @SpringBootTest(classes = csd.app.Main.class, webEnvironment = WebEnvironment.RANDOM_PORT)
@@ -62,9 +59,6 @@ class ProductIntegrationTest {
 
         @Autowired
         private UserInfoRepository userInfos;
-
-        @Autowired
-        private UserRecommendationRepository userRecommendations;
 
         @Autowired
         private BCryptPasswordEncoder encoder;
@@ -305,73 +299,4 @@ class ProductIntegrationTest {
                 assertEquals(400, result.getStatusCode().value());
                 assertEquals("Product not found: 999", root.path("message").asText());
         }        
-        
-        @Test
-        public void updateUserRecommendation_ValidRecommendation_Success() throws Exception {
-                User user = new User("tester2", "blabla@hotmail.com",
-                                encoder.encode("password1"));
-                users.save(user);
-                Long userId = user.getId();
-                UserInfo userInfo = new UserInfo(userId, user.getUsername(),
-                                "SINGAPORE01234567", 87231231);
-                userInfos.save(userInfo);
-                UserRecommendation userRecommendation = new UserRecommendation("NONE", userId);      
-                userRecommendations.save(userRecommendation);
-                RecommendationRequest recommendationRequest = new RecommendationRequest("ELECTRONICS", user.getUsername());
-                
-                URI uri = new URI(baseUrl + port + "/api/products/recommendation/update");
-
-                ResponseEntity<String> result = restTemplate.exchange(uri, HttpMethod.PUT,
-                                new HttpEntity<RecommendationRequest>(recommendationRequest), String.class);
-                JsonNode root = objectMapper.readTree(result.getBody());
-
-                UserRecommendation updatedUserRecommendation = userRecommendations.findById(userId).get();
-                assertEquals(200, result.getStatusCode().value());
-                assertEquals("User recommendation updated successfully", root.path("message").asText());
-                assertEquals(recommendationRequest.getRecommendation(), updatedUserRecommendation.getRecommendation());
-        }
-
-        @Test
-        public void updateUserRecommendation_InvalidRecommendation_Failure() throws Exception {
-                User user = new User("tester2", "blabla@hotmail.com",
-                                encoder.encode("password1"));
-                users.save(user);
-                Long userId = user.getId();
-                UserInfo userInfo = new UserInfo(userId, user.getUsername(),
-                                "SINGAPORE01234567", 87231231);
-                userInfos.save(userInfo);
-                UserRecommendation userRecommendation = new UserRecommendation("NONE", userId);      
-                userRecommendations.save(userRecommendation);
-                RecommendationRequest recommendationRequest = new RecommendationRequest("", user.getUsername());
-
-                URI uri = new URI(baseUrl + port + "/api/products/recommendation/update");
-
-                ResponseEntity<String> result = restTemplate.exchange(uri, HttpMethod.PUT,
-                                new HttpEntity<RecommendationRequest>(recommendationRequest), String.class);
-                JsonNode root = objectMapper.readTree(result.getBody());
-
-                assertEquals(400, result.getStatusCode().value());
-                assertEquals("Recommendation cannot be empty", root.path("message").asText());
-        }
-
-        @Test
-        public void getUserRecommendation_Success() throws Exception {
-                User user = new User("tester2", "blabla@hotmail.com",
-                                encoder.encode("password1"));
-                users.save(user);
-                Long userId = user.getId();
-                UserInfo userInfo = new UserInfo(userId, user.getUsername(),
-                                "SINGAPORE01234567", 87231231);
-                userInfos.save(userInfo);
-                UserRecommendation userRecommendation = new UserRecommendation("ELECTRONICS", userId);
-                userRecommendations.save(userRecommendation);
-
-                URI uri = new URI(baseUrl + port + "/api/products/recommendation/" + userId);
-
-                ResponseEntity<String> result = restTemplate.getForEntity(uri, String.class);
-                String validRecommendation = result.getBody();
-                assertNotNull(validRecommendation);
-                assertEquals(200, result.getStatusCode().value());
-                assertEquals(userRecommendation.getRecommendation(), validRecommendation);
-        }
 }
