+++ backend-spring-boot/src/test/java/csd/user/UserServiceTest.java
@@ -1,0 +1,44 @@
+// package csd.user;
+
+// import csd.app.user.User;
+// import csd.app.user.UserRepository;
+// import csd.app.user.UserInfo;
+// import csd.app.user.UserInfoRepository;
+// import csd.app.user.UserRecommendation;
+// import csd.app.user.UserRecommendationRepository;
+// import csd.app.user.UserServiceImpl;
+
+// import static org.junit.jupiter.api.Assertions.assertEquals;
+// import static org.junit.jupiter.api.Assertions.assertNotNull;
+// import static org.junit.jupiter.api.Assertions.assertNull;
+// import static org.junit.jupiter.api.Assertions.assertThrows;
+// import static org.mockito.ArgumentMatchers.any;
+// import static org.mockito.Mockito.when;
+
+// import org.junit.jupiter.api.Test;
+// import org.junit.jupiter.api.extension.ExtendWith;
+// import org.mockito.InjectMocks;
+// import org.mockito.Mock;
+// import org.mockito.junit.jupiter.MockitoExtension;
+// import org.springframework.security.crypto.password.PasswordEncoder;
+
+// @ExtendWith(MockitoExtension.class)
+// public class UserServiceTest {
+    
+//     @Mock
+//     private UserRepository users;
+
+//     @Mock
+//     private UserInfoRepository userInfos;
+
+//     @Mock
+//     private UserRecommendationRepository userRecommendations;
+
+//     @Mock
+//     PasswordEncoder encoder;
+
+//     @InjectMocks
+//     private UserServiceImpl userService;
+
+
+// }
